text
What this project does:
This project is created as a part of the Web Design Career programme from Training Dragon. It is a Mobile web design sample project created for a Nutritionist.

Why the project is useful
The project shows how to create webpages for mobile size, tablet size, laptop size and desktop size viewport.

How users can get started with the project
Start page is the index.html.
Where users can get help with your project
In case of any issue, post query on Git Hub.
Who maintains and contributes to the project
Payal Priyadarshini created this project.